+++ src/LibreSpot.Desktop/Views/CustomAppearanceDetailsSection.xaml.cs
@@ -4,9 +4,9 @@ using UserControl = System.Windows.Controls.UserControl;
 
 namespace LibreSpot.Desktop.Views;
 
-public partial class CustomAppearanceSection : UserControl
+public partial class CustomAppearanceDetailsSection : UserControl
 {
-    public CustomAppearanceSection()
+    public CustomAppearanceDetailsSection()
     {
         InitializeComponent();
     }